requirements.txt

flask
openai
supabase
line-bot-sdk==3.4.0
